refactor, fix pcntl_fork()
