+++ ext/pcntl/pcntl.c
@@ -50,6 +50,8 @@
 # define NSIG 32
 #endif
 
+#include "Zend/zend_timer.h"
+
 ZEND_DECLARE_MODULE_GLOBALS(pcntl)
 static PHP_GINIT_FUNCTION(pcntl);
 
@@ -531,6 +533,10 @@ PHP_FUNCTION(pcntl_fork)
 	if (id == -1) {
 		PCNTL_G(last_error) = errno;
 		php_error_docref(NULL, E_WARNING, "Error %d", errno);
+	} else if (id == 0) {
+		#ifdef ZEND_TIMER
+			zend_timer_create();
+		#endif
 	}
 
 	RETURN_LONG((zend_long) id);
